prepare for i386 implementation
