+++ CMakeLists.txt
@@ -6,9 +6,10 @@ set(CMAKE_CXX_COMPILER "/usr/bin/clang++")
 project(UlmerOS VERSION 0.1)
 enable_language(ASM)
 
-set(TARGET_ARCH "x86_64")
+set(TARGET_ARCH "i386")
 set(TARGET_PLATFORM "pc")
-set(ARCH_FLAGS "-mno-red-zone -mcmodel=kernel -mno-mmx -mno-sse -mno-avx")
+
+include(${CMAKE_CURRENT_SOURCE_DIR}/arch/${TARGET_ARCH}/arch.cmake)
 
 set(CMAKE_BUILD_TYPE Debug)
 if("${CMAKE_BUILD_TYPE}" STREQUAL "Debug")
