+++ arch/x86_64/arch.cmake
@@ -1,0 +1,3 @@
+set(ARCH_FLAGS "-mno-red-zone -mcmodel=kernel -mno-mmx -mno-sse -mno-avx")
+
+add_definitions(-Dx86_64 -DARCH=x86_64)
