+++ arch/x86_64/include/fpu.h
@@ -1,3 +1,4 @@
 #pragma once
 
 #include <x86/fpu.h>
+#include <x86/features.h>
