+++ arch/x86_common/atomic.cpp
