+++ arch/x86_common/include/fpu.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <features.h>
+#include <x86/features.h>
 
 struct FpuContextStruct;
